Add possibility to set name for downloaded files
This commit allows to rename the downloaded asset to the name that is
used as key.

That is triggered by setting the key "rename" to `true` for the element
like this:

"tool" : {
    "url" : "https://.../tool-v1.0.0.phar",
    "rename" : true
}

This will downloda the file tool-v1.0.0.phar and rename it to the key of the
entry, so in this case "tool". So it can be invoked using
"./vendor/bin/tool"